requirements.txt
# requires python version 3.7
numpy
matplotlib
gym
joblib
scipy
opencv-python
torch
tensorboard
munch
PyYAML

# for PPO, install these requirements
# stable-baselines==2.10.2
# tensorflow==1.15.0
# pip install gym==0.19.0
# pip install protobuf==3.20.*
